refactor(authd): create separate modules to handle each type of requests
