+++ safe-authd/authd.rs
@@ -10,11 +10,12 @@ use log::debug;
 use structopt::{self, StructOpt};
 
 use super::quic_client::quic_send;
+use super::requests::process_request;
 use super::update::update_commander;
 use daemonize::Daemonize;
 use failure::{Error, Fail, ResultExt};
-use futures::{Async, Future, Poll, Stream};
-use safe_api::{SafeAuthReq, SafeAuthenticator};
+use futures::{Future, Stream};
+use safe_api::SafeAuthenticator;
 use slog::{Drain, Logger};
 use std::collections::{BTreeMap, BTreeSet};
 use std::fs::File;
@@ -35,26 +36,23 @@ use url::Url;
 // Frequency for checking pensing auth requests
 const AUTH_REQS_CHECK_FREQ: u64 = 2000;
 
-// Maximum number of allowed auth reqs notifs subscriptors
-const MAX_NUMBER_OF_NOTIF_SUBSCRIPTIONS: usize = 3;
-
 #[derive(Debug, Clone)]
-struct AuthReq {
+pub struct AuthReq {
     pub app_id: String,
     pub tx: mpsc::Sender<bool>,
 }
 
 // List of authorisation requests indexed by their request id
-type AuthReqsList = BTreeMap<u32, AuthReq>;
+pub type AuthReqsList = BTreeMap<u32, AuthReq>;
 
 // A thread-safe queue to keep the list of authorisation requests
-type SharedAuthReqsHandle = Arc<Mutex<AuthReqsList>>;
+pub type SharedAuthReqsHandle = Arc<Mutex<AuthReqsList>>;
 
 // A thread-safe handle to keep the SafeAuthenticator instance
-type SharedSafeAuthenticatorHandle = Arc<Mutex<SafeAuthenticator>>;
+pub type SharedSafeAuthenticatorHandle = Arc<Mutex<SafeAuthenticator>>;
 
 // A thread-safe handle to keep the list of notifications subscriptors' endpoints
-type SharedNotifEndpointsHandle = Arc<Mutex<BTreeSet<String>>>;
+pub type SharedNotifEndpointsHandle = Arc<Mutex<BTreeSet<String>>>;
 
 const SAFE_AUTHD_PID_FILE: &str = "/tmp/safe-authd.pid";
 const SAFE_AUTHD_STDOUT_FILE: &str = "/tmp/safe-authd.out";
@@ -409,522 +407,6 @@ fn handle_request(
             .map(move |_| info!(log3, "Request complete"))
             .map_err(move |e| error!(log2, "Request Failed"; "reason" => %e.pretty())),
     )
-}
-
-fn process_request(
-    safe_auth_handle: SharedSafeAuthenticatorHandle,
-    auth_reqs_handle: SharedAuthReqsHandle,
-    notif_endpoints_handle: SharedNotifEndpointsHandle,
-    req: &[u8],
-) -> ProcessRequest {
-    ProcessRequest::HandleRequest {
-        safe_auth_handle,
-        auth_reqs_handle,
-        notif_endpoints_handle,
-        req: req.to_vec(),
-    }
-}
-
-enum ProcessRequest {
-    HandleRequest {
-        safe_auth_handle: SharedSafeAuthenticatorHandle,
-        auth_reqs_handle: SharedAuthReqsHandle,
-        notif_endpoints_handle: SharedNotifEndpointsHandle,
-        req: Vec<u8>,
-    },
-    ProcessingResponse {
-        safe_auth_handle: SharedSafeAuthenticatorHandle,
-        auth_reqs_handle: SharedAuthReqsHandle,
-        rx: mpsc::Receiver<bool>,
-        req_id: u32,
-        auth_req_str: String,
-    },
-}
-
-impl Future for ProcessRequest {
-    type Item = Box<[u8]>;
-    type Error = Error;
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        use self::ProcessRequest::*;
-
-        // TODO: implement JSON-RPC rather.
-        // Temporarily prefix message with "[AUTHD_ERROR]" to signal error to the caller,
-        // once we have JSON-RPC we can adhere to its format for errors.
-        let future_err = |str: String| -> Poll<Self::Item, Self::Error> {
-            Ok(Async::Ready(
-                format!("[AUTHD_ERROR]:SAFE Authenticator: {}", str)
-                    .into_bytes()
-                    .into(),
-            ))
-        };
-
-        let future_ok = |str: String| -> Poll<Self::Item, Self::Error> {
-            Ok(Async::Ready(str.into_bytes().into()))
-        };
-
-        loop {
-            match self {
-                ProcessingResponse {
-                    safe_auth_handle,
-                    auth_reqs_handle,
-                    rx,
-                    req_id,
-                    auth_req_str,
-                } => {
-                    match rx.poll() {
-                        Ok(Async::Ready(Some(is_allowed))) => {
-                            rx.close();
-                            if is_allowed {
-                                let safe_auth: &mut SafeAuthenticator =
-                                    &mut *(safe_auth_handle.lock().unwrap());
-                                match safe_auth.authorise_app(auth_req_str) {
-                                    Ok(resp) => {
-                                        println!("Authorisation request ({}) was allowed and response sent back to the application", req_id);
-                                        return future_ok(resp);
-                                    }
-                                    Err(err) => {
-                                        println!("Failed to authorise application: {}", err);
-                                        return future_err(err.to_string());
-                                    }
-                                }
-                            } else {
-                                let msg = format!("Authorisation request ({}) was denied", req_id);
-                                println!("{}", msg);
-                                return future_err(msg);
-                            }
-                        }
-                        Ok(Async::NotReady) => {
-                            return Ok(Async::NotReady);
-                        }
-                        Ok(Async::Ready(None)) | Err(_) => {
-                            rx.close();
-                            // We didn't get a response in a timely manner, we cannot allow the list
-                            // to grow infinitelly, so let's remove the request from it
-                            let auth_reqs_list: &mut AuthReqsList =
-                                &mut *(auth_reqs_handle.lock().unwrap());
-                            auth_reqs_list.remove(&req_id);
-                            let msg = "Failed to get authorision response";
-                            println!("{}", msg);
-                            return future_err(msg.to_string());
-                        }
-                    }
-                }
-                HandleRequest {
-                    safe_auth_handle,
-                    auth_reqs_handle,
-                    notif_endpoints_handle,
-                    req,
-                } => {
-                    if req.len() < 4 || &req[0..4] != b"GET " {
-                        return future_err("Missing GET".to_string());
-                    }
-                    if req[4..].len() < 2 || &req[req.len() - 2..] != b"\r\n" {
-                        return future_err("Missing \\r\\n".to_string());
-                    }
-                    let req = &req[4..req.len() - 2];
-                    let end = req
-                        .iter()
-                        .position(|&c| c == b' ')
-                        .unwrap_or_else(|| req.len());
-                    let path = match str::from_utf8(&req[..end]).context("Path is malformed UTF-8")
-                    {
-                        Ok(path) => path,
-                        Err(err) => return future_err(err.to_string()),
-                    };
-                    let req_args: Vec<&str> = path.split('/').collect();
-
-                    let safe_auth_handle = safe_auth_handle.clone();
-                    let safe_auth: &mut SafeAuthenticator =
-                        &mut *(safe_auth_handle.lock().unwrap());
-
-                    println!("Processing new incoming request: '{}'", req_args[1]);
-                    match req_args[1] {
-                        "login" => {
-                            if req_args.len() != 4 {
-                                return future_err(
-                                    "Incorrect number of arguments for 'login' action".to_string(),
-                                );
-                            } else {
-                                println!("Logging in to SAFE account...");
-                                let secret = req_args[2];
-                                let password = req_args[3];
-
-                                match safe_auth.log_in(secret, password) {
-                                    Ok(_) => {
-                                        let msg = "Logged in successfully!";
-                                        println!("{}", msg);
-                                        return future_ok(msg.to_string());
-                                    }
-                                    Err(err) => {
-                                        let msg = format!(
-                                            "Error occurred when trying to log in: {}",
-                                            err
-                                        );
-                                        println!("{}", msg);
-                                        return future_err(err.to_string());
-                                    }
-                                }
-                            }
-                        }
-                        "logout" => {
-                            if req_args.len() != 2 {
-                                return future_err(
-                                    "Incorrect number of arguments for 'logout' action".to_string(),
-                                );
-                            } else {
-                                println!("Logging out...");
-                                match safe_auth.log_out() {
-                                    Ok(()) => {
-                                        let msg = "Logged out successfully";
-                                        println!("{}", msg);
-                                        return future_ok(msg.to_string());
-                                    }
-                                    Err(err) => {
-                                        let msg = format!("Failed to log out: {}", err);
-                                        println!("{}", msg);
-                                        return future_err(msg);
-                                    }
-                                }
-                            }
-                        }
-                        "create" => {
-                            if req_args.len() != 5 {
-                                return future_err(
-                                    "Incorrect number of arguments for 'create' action".to_string(),
-                                );
-                            } else {
-                                println!("Creating an account in SAFE...");
-                                let secret = req_args[2];
-                                let password = req_args[3];
-                                let sk = req_args[4];
-
-                                match safe_auth.create_acc(sk, secret, password) {
-                                    Ok(_) => {
-                                        let msg = "Account created successfully";
-                                        println!("{}", msg);
-                                        return future_ok(msg.to_string());
-                                    }
-                                    Err(err) => {
-                                        println!(
-                                            "Error occurred when trying to create SAFE account: {}",
-                                            err
-                                        );
-                                        return future_err(err.to_string());
-                                    }
-                                }
-                            }
-                        }
-                        "authorise" => {
-                            if req_args.len() != 3 {
-                                return future_err(
-                                    "Incorrect number of arguments for 'authorise' action"
-                                        .to_string(),
-                                );
-                            } else {
-                                println!("Authorising application...");
-                                // TODO: automatically reject if there are too many pending auth reqs
-                                let auth_req_str = req_args[2];
-                                match safe_auth.decode_req(auth_req_str) {
-                                    Ok((req_id, safe_auth_req)) => {
-                                        println!(
-                                            "Sending request to user to allow/deny request..."
-                                        );
-
-                                        let rx = enqueue_auth_req(
-                                            req_id,
-                                            safe_auth_req,
-                                            auth_reqs_handle,
-                                        );
-
-                                        *self = ProcessingResponse {
-                                            safe_auth_handle: safe_auth_handle.clone(),
-                                            auth_reqs_handle: auth_reqs_handle.clone(),
-                                            rx,
-                                            req_id,
-                                            auth_req_str: auth_req_str.to_string(),
-                                        };
-                                    }
-                                    Err(err) => {
-                                        println!("{}", err);
-                                        return future_err(err.to_string());
-                                    }
-                                }
-                            }
-                        }
-                        "authed-apps" => {
-                            if req_args.len() != 2 {
-                                return future_err(
-                                    "Incorrect number of arguments for 'authed-apps' action"
-                                        .to_string(),
-                                );
-                            } else {
-                                println!("Obtaining list of authorised applications...");
-                                match safe_auth.authed_apps() {
-                                    Ok(resp) => {
-                                        println!("List of authorised apps sent");
-                                        return future_ok(format!("{:?}", resp));
-                                    }
-                                    Err(err) => {
-                                        println!("Failed to get list of authorised apps: {}", err);
-                                        return future_err(err.to_string());
-                                    }
-                                }
-                            }
-                        }
-                        "revoke" => {
-                            if req_args.len() != 3 {
-                                return future_err(
-                                    "Incorrect number of arguments for 'revoke' action".to_string(),
-                                );
-                            } else {
-                                println!("Revoking application...");
-                                let app_id = req_args[2];
-
-                                match safe_auth.revoke_app(app_id) {
-                                    Ok(()) => {
-                                        let msg = "Application revoked successfully";
-                                        println!("{}", msg);
-                                        return future_ok(msg.to_string());
-                                    }
-                                    Err(err) => {
-                                        println!(
-                                            "Failed to revoke application '{}': {}",
-                                            app_id, err
-                                        );
-                                        return future_err(err.to_string());
-                                    }
-                                }
-                            }
-                        }
-                        "auth-reqs" => {
-                            if req_args.len() != 2 {
-                                return future_err(
-                                    "Incorrect number of arguments for 'auth-reqs' action"
-                                        .to_string(),
-                                );
-                            } else {
-                                println!("Obtaining list of pending authorisation requests...");
-                                let auth_reqs_list: &mut AuthReqsList =
-                                    &mut *(auth_reqs_handle.lock().unwrap());
-                                let resp: BTreeSet<String> = auth_reqs_list
-                                    .iter()
-                                    .map(|(req_id, auth_req)| {
-                                        format!("Req ID: {} - App ID: {}", req_id, auth_req.app_id)
-                                    })
-                                    .collect();
-
-                                println!("List of pending authorisation requests sent");
-                                return future_ok(format!("{:?}", resp));
-                            }
-                        }
-                        "allow" => {
-                            if req_args.len() != 3 {
-                                return future_err(
-                                    "Incorrect number of arguments for 'allow' action".to_string(),
-                                );
-                            } else {
-                                println!("Allowing authorisation request...");
-                                let auth_req_id = req_args[2];
-                                let auth_reqs_list: &mut AuthReqsList =
-                                    &mut *(auth_reqs_handle.lock().unwrap());
-                                let req_id = match auth_req_id.parse::<u32>() {
-                                    Ok(id) => id,
-                                    Err(err) => return future_err(err.to_string()),
-                                };
-                                match auth_reqs_list.remove(&req_id) {
-                                    Some(mut auth_req) => match auth_req.tx.try_send(true) {
-                                        Ok(_) => {
-                                            let msg = format!(
-                                                "Authorisation request ({}) allowed successfully",
-                                                auth_req_id
-                                            );
-                                            println!("{}", msg);
-                                            return future_ok(msg);
-                                        }
-                                        Err(_) => {
-                                            let msg = format!("Failed to allow authorisation request '{}' since the response couldn't be sent to the requesting application", auth_req_id);
-                                            println!("{}", msg);
-                                            return future_err(msg);
-                                        }
-                                    },
-                                    None => {
-                                        let msg = format!(
-                                            "No pending authorisation request found with id '{}'",
-                                            auth_req_id
-                                        );
-                                        println!("{}", msg);
-                                        return future_err(msg);
-                                    }
-                                }
-                            }
-                        }
-                        "deny" => {
-                            if req_args.len() != 3 {
-                                return future_err(
-                                    "Incorrect number of arguments for 'deny' action".to_string(),
-                                );
-                            } else {
-                                println!("Denying authorisation request...");
-                                let auth_req_id = req_args[2];
-                                let auth_reqs_list: &mut AuthReqsList =
-                                    &mut *(auth_reqs_handle.lock().unwrap());
-                                let req_id = match auth_req_id.parse::<u32>() {
-                                    Ok(id) => id,
-                                    Err(err) => return future_err(err.to_string()),
-                                };
-                                match auth_reqs_list.remove(&req_id) {
-                                    Some(mut auth_req) => match auth_req.tx.try_send(false) {
-                                        Ok(_) => {
-                                            let msg = format!(
-                                                "Authorisation request ({}) denied successfully",
-                                                auth_req_id
-                                            );
-                                            println!("{}", msg);
-                                            return future_ok(msg);
-                                        }
-                                        Err(_) => {
-                                            let msg = format!("Authorisation request '{}' was already denied since the response couldn't be sent to the requesting application", auth_req_id);
-                                            println!("{}", msg);
-                                            return future_err(msg);
-                                        }
-                                    },
-                                    None => {
-                                        let msg = format!(
-                                            "No pending authorisation request found with id '{}'",
-                                            auth_req_id
-                                        );
-                                        println!("{}", msg);
-                                        return future_err(msg);
-                                    }
-                                }
-                            }
-                        }
-                        "subscribe" => {
-                            if req_args.len() != 3 {
-                                return future_err(
-                                    "Incorrect number of arguments for 'subscribe' action"
-                                        .to_string(),
-                                );
-                            } else {
-                                println!("Subscribing to authorisation requests notifications...");
-                                let mut notif_endpoint = match urlencoding::decode(req_args[2]) {
-                                    Ok(url) => url,
-                                    Err(err) => {
-                                        let msg = format!(
-                                        "Subscription rejected, the endpoint URL ('{}') is invalid: {:?}",
-                                        req_args[2], err
-                                    );
-                                        println!("{}", msg);
-                                        return future_err(msg);
-                                    }
-                                };
-                                let notif_endpoints_list: &mut BTreeSet<String> =
-                                    &mut *(notif_endpoints_handle.lock().unwrap());
-                                if notif_endpoints_list.len() >= MAX_NUMBER_OF_NOTIF_SUBSCRIPTIONS {
-                                    let msg = format!("Subscription rejected. Maximum number of subscriptions ({}) has been already reached", MAX_NUMBER_OF_NOTIF_SUBSCRIPTIONS);
-                                    println!("{}", msg);
-                                    return future_err(msg);
-                                } else {
-                                    if notif_endpoint.ends_with('/') {
-                                        notif_endpoint.pop();
-                                    }
-                                    notif_endpoints_list.insert(notif_endpoint.clone());
-                                    let msg = format!(
-                                    "Subscription successful. Endpoint '{}' will receive authorisation requests notifications",
-                                    notif_endpoint
-                                );
-                                    println!("{}", msg);
-                                    return future_ok(msg);
-                                }
-                            }
-                        }
-                        "unsubscribe" => {
-                            if req_args.len() != 3 {
-                                return future_err(
-                                    "Incorrect number of arguments for 'unsubscribe' action"
-                                        .to_string(),
-                                );
-                            } else {
-                                println!(
-                                    "Unsubscribing from authorisation requests notifications..."
-                                );
-                                let notif_endpoint = match urlencoding::decode(req_args[2]) {
-                                    Ok(url) => url,
-                                    Err(err) => {
-                                        let msg = format!(
-                                        "Unsubscription request rejected, the endpoint URL ('{}') is invalid: {:?}",
-                                        req_args[2], err
-                                    );
-                                        println!("{}", msg);
-                                        return future_err(msg);
-                                    }
-                                };
-                                let notif_endpoints_list: &mut BTreeSet<String> =
-                                    &mut *(notif_endpoints_handle.lock().unwrap());
-
-                                if notif_endpoints_list.remove(&notif_endpoint) {
-                                    let msg = format!(
-                                    "Unsubscription successful. Endpoint '{}' will no longer receive authorisation requests notifications",
-                                    notif_endpoint
-                                );
-                                    println!("{}", msg);
-                                    return future_ok(msg);
-                                } else {
-                                    let msg = format!(
-                                    "Unsubscription request ignored, no such the endpoint URL ('{}') was found to be subscribed",
-                                    notif_endpoint
-                                );
-                                    println!("{}", msg);
-                                    return future_err(msg);
-                                }
-                            }
-                        }
-                        other => {
-                            println!(
-                                "Action '{}' not supported or unknown by the Authenticator daemon",
-                                other
-                            );
-                            return future_err("Action not supported or unknown".to_string());
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-fn enqueue_auth_req(
-    req_id: u32,
-    req: SafeAuthReq,
-    auth_reqs_handle: &SharedAuthReqsHandle,
-) -> mpsc::Receiver<bool> {
-    let (tx, rx): (mpsc::Sender<bool>, mpsc::Receiver<bool>) = mpsc::channel(32);
-    match req {
-        SafeAuthReq::Auth(app_auth_req) => {
-            println!("The following application authorisation request was received:");
-            println!("{:?}", app_auth_req);
-
-            // Let's add it to the list of pending authorisation requests
-            let auth_req = AuthReq {
-                app_id: app_auth_req.app.id,
-                tx,
-            };
-            let auth_reqs_list: &mut AuthReqsList = &mut *(auth_reqs_handle.lock().unwrap());
-            auth_reqs_list.insert(req_id, auth_req);
-        }
-        SafeAuthReq::Containers(cont_req) => {
-            println!("The following authorisation request for containers was received:");
-            println!("{:?}", cont_req);
-        }
-        SafeAuthReq::ShareMData(share_mdata_req) => {
-            println!("The following authorisation request to share a MutableData was received:");
-            println!("{:?}", share_mdata_req);
-        }
-        SafeAuthReq::Unregistered(_) => {
-            // we simply allow unregistered authorisation requests
-        }
-    }
-    rx
 }
 
 fn monitor_pending_auth_reqs(
