+++ safe-authd/requests/unsubscribe.rs
@@ -1,0 +1,50 @@
+// Copyright 2019 MaidSafe.net limited.
+//
+// This SAFE Network Software is licensed to you under The General Public License (GPL), version 3.
+// Unless required by applicable law or agreed to in writing, the SAFE Network Software distributed
+// under the GPL Licence is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied. Please review the Licences for the specific language governing
+// permissions and limitations relating to use of the SAFE Network Software.
+
+use crate::authd::SharedNotifEndpointsHandle;
+use std::collections::BTreeSet;
+
+pub fn process_req(
+    args: &[&str],
+    notif_endpoints_handle: SharedNotifEndpointsHandle,
+) -> Result<String, String> {
+    if args.len() != 1 {
+        Err("Incorrect number of arguments for 'unsubscribe' action".to_string())
+    } else {
+        println!("Unsubscribing from authorisation requests notifications...");
+        let notif_endpoint = match urlencoding::decode(args[0]) {
+            Ok(url) => url,
+            Err(err) => {
+                let msg = format!(
+                    "Unsubscription request rejected, the endpoint URL ('{}') is invalid: {:?}",
+                    args[0], err
+                );
+                println!("{}", msg);
+                return Err(msg);
+            }
+        };
+        let notif_endpoints_list: &mut BTreeSet<String> =
+            &mut *(notif_endpoints_handle.lock().unwrap());
+
+        if notif_endpoints_list.remove(&notif_endpoint) {
+            let msg = format!(
+                "Unsubscription successful. Endpoint '{}' will no longer receive authorisation requests notifications",
+                notif_endpoint
+            );
+            println!("{}", msg);
+            Ok(msg)
+        } else {
+            let msg = format!(
+                "Unsubscription request ignored, no such the endpoint URL ('{}') was found to be subscribed",
+                notif_endpoint
+            );
+            println!("{}", msg);
+            Err(msg)
+        }
+    }
+}
